backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.9.0
python-dotenv>=1.0.1
openai>=1.10.0
supabase>=2.3.0
httpx>=0.26.0
razorpay>=1.3.0